+++ launch-daemon.sh
@@ -1,5 +1,5 @@
 #!/bin/bash
 if [[ ! $(pgrep cartel-daemon) ]]; then
-    (cartel-daemon >/dev/null &)
+    cartel-daemon --detach >/dev/null
 fi